Added flamegraph.pl + improved dart support

+++ timemory/common.py
@@ -61,6 +61,7 @@ __all__ = [
     "dart_measurement",
     "dart_measurement_file",
     "write_ctest_notes",
+    "get_bin_script",
 ]
 
 CO_GENERATOR = 0x0020
@@ -168,8 +169,8 @@ else:
         return False
 
 
-def popen(cmd, outf=None, keep_going=True, timeout=None):
-    """Execute a command"""
+def popen(cmd, outf=None, keep_going=True, timeout=None, shell=False):
+    """Execute a command and write to a file or get the output"""
 
     def handle_error(ret, cmd, keep_going):
         """Handle error messaging for executed command"""
@@ -192,39 +193,62 @@ def popen(cmd, outf=None, keep_going=True, timeout=None):
 
     import subprocess as sp
 
-    p = sp.Popen(cmd)
+    p = None
     outs = None
     errs = None
-    try:
-        if timeout is not None:
-            outs, errs = p.communicate(timeout=timeout)
+    if shell:
+        p = sp.run(
+            " ".join(cmd),
+            shell=True,
+            check=False,
+            capture_output=True,
+        )
+    else:
+        p = sp.Popen(cmd)
+        try:
+            if timeout is not None:
+                outs, errs = p.communicate(timeout=timeout)
+            else:
+                outs, errs = p.communicate()
+        except sp.TimeoutExpired:
+            p.kill()
+
+    if p.returncode != 0:
+        if errs is not None:
+            print("{}".format(errs.decode("utf-8")))
         else:
-            outs, errs = p.communicate()
-    except sp.TimeoutExpired:
-        p.kill()
+            print("{}".format(p.stderr.decode("utf-8")))
 
     handle_error(p.returncode, cmd, keep_going)
 
-    if errs is not None:
-        print("{}".format(outs.decode("utf-8")))
+    if outf is not None:
+        with open(outf, "w") as f:
+            if outs is not None:
+                f.write(outs)
+            else:
+                f.write(p.stdout)
 
-    if outs is not None and outf is not None:
-        f = open(outf, "w")
-        f.write(outs)
+    if outs is None:
+        outs = "{}".format(p.stdout.decode("utf-8"))
+
+    if errs is None:
+        errs = "{}".format(p.stderr.decode("utf-8"))
+
+    return (outs, errs)
 
 
 def dart_measurement(name, value):
     """Prints out an XML tag which gets detected by CTest and recorded by CDash"""
     print(
-        f'<DartMeasurement name="{name}" type="numeric/double">{value}</DartMeasurementFile>'
+        f'<DartMeasurement name="{name}" type="numeric/double">{value}</DartMeasurement>'
     )
 
 
-def dart_measurement_file(name, path, format="png"):
+def dart_measurement_file(name, path, format="png", type="image"):
     """Prints out an XML tag which gets detected by CTest and uploaded to CDash"""
     _path = os.path.abspath(path)
     print(
-        f'<DartMeasurementFile name="{name}" type="image/{format}">{_path}</DartMeasurementFile>'
+        f'<DartMeasurementFile name="{name}" type="{type}/{format}">{_path}</DartMeasurementFile>'
     )
 
 
@@ -238,3 +262,17 @@ def write_ctest_notes(fname, path=None, mode="a"):
     fname = os.path.abspath(fname)
     with open(os.path.join(path, "CTestNotes.txt"), "a") as ofs:
         ofs.write(f'\nlist(APPEND CTEST_NOTES_FILES "{fname}")\n')
+
+
+def get_bin_script(fname):
+    """ Returns the path to a script in the internal bin folder """
+    _this_dir = os.path.dirname(__file__)
+    _fpath = os.path.join(_this_dir, "bin", fname)
+    if os.path.exists(_fpath):
+        return _fpath
+    import glob
+
+    for itr in glob.glob(f"{_fpath}*"):
+        if os.path.exists(itr):
+            return itr
+    return None
